refactor: improved type definition
